aave-avax-dashboard/backend/requirements.txt
fastapi
uvicorn
python-dotenv
web3
eth-defi[aave,uniswap]
